Removed BpkButtonWithIcon.kt
